elastic: Add traced HTTPClient, tests for v3 and v5.
TracedHTTPClient hooks into the http.Transport to trace requests and
capture spans for Elasticsearch. It's generic enough that it could be
used for other Elasticsearch clients that allow you to override the http
client.

The tests and example show the pattern for tracing the elastic client,
specifically using elastic.v5. Older versions do not support calling
`Do` with a context so they won't apply. But in theory this should be
compatible with new version going forward.

docker-compose is updated to run ES for the integration tests.