Remove tlsConfigRegister in MySQL DSN parser
If you do include a custom TLS config in the MySQL DSN, this code will return an error because it hasn't been registered in this particular tlsConfigRegister (and there isn't even a function available for doing so), even if you correctly register with the go-sql-driver/mysql driver. Since the TLS config isn't used in any of the trace metadata, we can just remove this bit.