contrib: move contributions to their own repository
This change moves the contributions (3rd party integrations) to their
own repository. This has been done by simply copying them over and
freezing dependencies on the root level, as well as freezing the gRPC's
v1.2's dependency in its own library.

As a result, the build system for the tracer is greatly simplified and
rake removed.