ddtrace/opentracer: s/Start/New and return tracer
This change alters the instantiation of the Opentracing method to one
that is more in line with other implementations in the wild, meaning
that now the call is named `New` and it returns the
`opentracing.Tracer` which can be used with
`opentracing.SetGlobalTracer` as per the official documentation.

This is more useful when using libraries such as `opentracing-contrib`
where the tracer is needed, and could potentially reduce some calls to
`opentracing.GlobalTracer`.