ddtrace/tracer: trace should use any available tracer.
Formerly, it was possible that a tracer would be stopped, but the trace in a
span's context would continue holding a reference to it. This would
cause that trace to try and push itself onto a stopped tracer.

Now, traces use the global tracer, and even if it wasn't the tracer that
created the spans, they will still flush, given that the global tracer
is a valid tracer.