ddtrace/tracer: prevent trace from acknowledging more spans after cap is (#223)
reached

This change ensures that once the span buffer reaches capacity, we will
not be handling nor flushing the trace anymore since we can no longer
keep track of the spans that are part of it. Otherwise, the flush
condition is met prematurely when not all spans in the buffer are
completed and we are open for a race condition: modifying a span while
flushing.