ddtrace/tracer: improve SpanFromContext API
This change provides for a safer API to SpanFromContext which assures
that the value `nil` is not returned anymore. This was confusing
and could potentially crash an app where one would attempt to call
methods on the span without formerly checking its non-nil value.

Now, when a span is not found, the second return value will be `false`
and the returned span will be a no-op, allowing users to call methods on
it without having to explicitly check that the span was found in the
context, if there is no interest to, while at the same time providing
the option to do so.