docs: improve documentation and add guidelines.
This change improves the README so that the documentation section comes
before the testing one, since it has higher importance.

It also moves the contributing guildelines into their own file
(CONTRIBUTING.md) as this is standard for GitHub and will cause PRs to
be preceeded with a warning that this repository has contribution
guidelines, linking to this file.