all: standardize span types
This change adds a set of `SpanType*` constants to the `ext` package
outlining all the possible span types that are currently used or can be
used by the agent. It also updates the entire codebase to make use of
these constants wherever possible.

The remaining `AppType*` constants are being discouraged from usage and
should be used only to categorize traces in the UI when no better
alternative exists.