ddtrace/tracer: priority sampling on by default
This is to improve the behavior that users observe when collecting
distributed traces.
With a default to on, traces that are being sampled will have the
parts produced by dd-trace-go submitted to the agent.

Currently, no option is provided to disable priority sampling, but code
for the original behavior has been kept. This may be reconsidered in the
future.