ddtrace/tracer: fix a race condition happening with asynchronous flushes.
This change fixes a race condition that would happen on rare occasions
when using the priority sampler and flushing asynchronous child spans of
a trace that would finish later than the root, usually at the same time
when the root was being flushed.

This fix also means that after the root of a trace has finished,
sampling priority may no longer be set.