circleci: vendor go-redis/redis for testing.
The redis package has new development on their master branch for a v7
release, and there are significant API changes. The current contrib may
need a rewrite once that API is stable and released. In the meantime,
this pins the version of redis that we test against, while still making
sure users can choose their own version.