contrib/net/http: add RTWithServiceName option for RoundTripper
While ServeMux and WrapHandler take a WithServiceName option,
WrapRoundTripper and WrapClient don't have something comparable.
It is possible to add a WithBefore option that sets the service name, but
that is not an ideal situation.

This commit adds the RTWithServiceName function, which will set the
service name for the RoundTripper.