contrib: specify the need for the major version in the README
It currently happens that libraries which we are integrating with are
releasing new major versions that we don't support. As a result, we need
to create new packages to support them. This implies not only creating a new
folder using the current convention, but also adding the version to the
path.

Since we've not done this until today, users will be tempted to use the
"versionless" paths for latest releases, which will be inaccurate.

This change adds a new convention to *always* specify the major version
number as a suffix to the path. This is also in line with the requirements
for modules.