ddtrace/tracer: fix regression when flushing based on size
During a previous change released as part of v1.20.0 in PR #498, the
flush channel was accidentally made blocking. This would cause a
deadlock on occasion when flushing after the payload was full.

Tests failed to catch this problem because they were not using the
constructor and instead had their own, which was made correctly. This
change also addresses that problem and removes the constructor from the
tests, now correctly reproducing the problem.