contrib/gin-gonic/gin: correct resource name
There was a mistake with setting the resource name for this integration.
We previously had used `(*gin.Context).HandlerName` to set the resource
which was wrong and returns the name of the function (e.g.
`main.main.func1`).

This change ensures that the resource name is set correctly, as in all
our HTTP integrations to method and path (e.g. "GET /path")