ddtrace/tracer: add service version to tracer. (#607)
This commit adds the concept of a service version to the tracer. The service version should be the current version of the application (service). the service version will be picked up from the DD_VERSION environment variable, or can be set with the WithServiceVersion StartOption when starting the tracer. WithServiceVersion takes priority over the environment variable.

Spans that have a service name equal to the application's service name (set with DD_SERVICE or the WithServiceName StartOption) will be given a "version" tag with the service version as its value.