contrib: update non-client integrations to use configured service name
Integrations set their own service name regardless of what is configured in
the tracer. This commit updates non-client integrations to pick up the
tracer's service name if one has been configured, and falls back to the
previous default service name for the integration. Integrations can still
be individually configured with their own service names which will override
the one configured in the tracer.

In order to do this, the tracer exposes its service name through the
internal/globalconfig package.