contrib/Shopify/sarama: use sarama.MinVersion in tests. (#761)
Starting in github.com/Shopify/sarama@v1.27.1, the default kafka version
that the client uses is 1.0.0 (sarama.V1_0_0_0) rather that the previous
0.8.2 (sarama.V0_8_2_0) Unfortunately, the Sarama client for Kafka 1.0.0
does not work with the mock broker, so this change breaks our tests. To
repair the situation, we specify that the client should use version 0.8.2.
If Sarama fixes their mock client in the future, we can use the default
version instead.