contrib/Shopify/sarama: Fix WrapAsyncProducer tracing support (#738)
Before a message is successfully delivered to Kafka, the Offset and
Partition are not valid, meaning they can't be used reliably as a span key
nor as tags.  Tracing the Async Producer also does not work at all with
kafka versions before 0.11 since they do not support the metadata header.

This commit fixes WrapAsyncProducer by using the spanID as a key when
tracking spans and by warning when older kafka versions are used.

Sadly fixing TestAsyncProducer tests is not possible because
sarama.MockBroker does not work for newer kafka versions. (See:
https://github.com/Shopify/sarama/issues/1665) This fix has been manually
verified and this test has been marked as skipped until the MockBroker is
fixed in sarama.

Fixes #483