contrib/gin-gonic/gin: support globalconfig.ServiceName()
This change adds support for `globalconfig.ServiceName()` to the gin
trace middleware. Because this middleware does not follow the same pattern
as other http frameworks and requires a service parameter to be passed in
we use the empty string value as an indicator that the default service
name should be used.

Fixes #767