profiler: Implement WithUploadTimeout
Also add support for the DD_PROFILING_UPLOAD_TIMEOUT used by other
DD profilers such as Java.

The timeout is implemented using a request context, which allows
removing the timeout option from the defaultClient and WithUDS option.

defaultHTTPTimeout is promoted to DefaultUploadTimeout to be consistent
with the other default options.

The order of preference is:
1. DefaultUploadTimeout
2. DD_PROFILING_UPLOAD_TIMEOUT
3. WithUploadTimeout()

Implements PROF-2810