README: Fix go get instructions (#913)
Fixes #911

The grpc sub-pkg not compiling might still cause issues for some use
cases, but they are mostly related to local development. For our users
it's good enough to steer them away from the problem for now. But I'm
planning to come up with a solution for grpc as well.