profiler: Fix infinite loop in maxPauseNs
I noticed this issue when running the TestProfilerInternal test by
itself, which consistently produced the following error:

```
$ go test ./profiler -run TestProfilerInternal
--- FAIL: TestProfilerInternal (0.20s)
    --- FAIL: TestProfilerInternal/collect (0.20s)
        profiler_test.go:189: missing batch
FAIL
FAIL	gopkg.in/DataDog/dd-trace-go.v1/profiler	0.395s
```

The reason for this timeout is that the loop inside of maxPauseNs was
not hitting its termination conditions when stats.NumGC is 0. The reason
for this is subtle:

- i is uint32, so it will always be <= 0
- periodStart is "1729-02-04" when running the metrics profile for the first
  time. This is caused by another issue which will be addressed in a
  follow-up patch.

As far as I can tell this issue should not occur in the real world
because the periodStart will not be 1729 after `profiler.run` resets the
metrics collectedAt timestamp.