contrib/confluentinc/confluent-kafka-go/kafka: Use kafka headers as context when producing
Before this patch when producing a message, the general producer context was
used when producing messages.

This made it impossible to attach a parent span to a produced message.

With this patch messages headers are checked when producing messages for a span
context and the span is created as a child if a parent span is present.

This patch deprecates the WithContext() option for consumer and producers.

Fixed #941