appsec/dyngo: add an instrumentation gateway draft
This draft instrumentation gateway allows to listen and react to specific
operation stacks. In this first attempt, it tries to avoid maps of operation
attributes and rather relies on Go structures. This allows to simplify the
event listener and emitter API by relying on the operation arguments/results
structure type. It also allows have strongly types lists of operation
arguments and results.

The main limitation comes from the absence of Go generics for now, that
implies the use of empty interface{} values along with some calls to the
reflection package.