appsec/waf: simply the WAF event listener
By enforcing the order of the WAF listener so that it properly comes after the data events,
it is then possible to only rely on them and have a cleaner callback.

This shows that a better abstraction should be developped in the future so that no such manual
hacks in the listener order is required.