dyngo: add run time type-checking of operations and listeners
An operation must now be pre-declared using `dyngo.RegisterOperation()`. It
allows to check if an operation is properly used at run time by checking
its argument and result types are correct. A few mistakes in the tests were
already caught thanks to it.

Event listeners are also checked and should of the form
`func (*dyngo.Operation, T)` or the newly accepted type `func (T)` which
is common for "leaf listeners" that don't need to have access to the
operation where the event happens.

The last important change is that it is now required to register an operation
in order to allow to listen to it. This makes possible to check the types of
the operation argument and result types used in the listener: a start event
listener should expect operation arguments, while a finish event listener
should expect operation results. This is now checked at run time.

For now, those errors still panic to keep the API simple to use and because
all of these errors would happen in case of development mistakes. The only
case that could happen at run time is in case of an event registration
happening before the operation's. If this becomes a problem in the future,
we could expose a new function allowing to know if an operation is registered
or not.