dyngo: port event listeners to static types
Event listener functions are now ported to `func (op *Operation, value interface{})`.
The value type that was previously used to register the event is now provided as first
argument of event registration functions (OnStart, OnFinish, OnData and Register).

Benchmarks shown the impact of the value types escaping to the heap when passed as
interface values. To avoid those extra and useless allocations, value types must
now be passed as nil pointers, such as (*T)(nil).

The resulting API is more difficult to use than the previous one, and helper functions
should be provided to enforce the overall type-safety and make it simpler and clearer
to use. Code generation (with go generate) should be used in the future if this API
survives over time.