appsec/waf: add libddwaf v1.0.12 bindings
No longer use go-libsqreen but the new libddwaf. The go-libsqreen
bindings were updated to the new API and a lot of
tests were added to better cover the corner cases.

The currently supported Go compilation targets are linux/amd64 and
darwin/amd64. Non supported targets still compiled
thanks to no-op stubs.

It is also possible to disable it either setting the env var
CGO_ENABLED=0 or the build tag without_ddwaf during
the program's compilation.

The Health() function is in charge of returning the WAF version along
with an error if the WAF got disabled at compilation time with an error
explaining why.