.circleci: Change cache key to actually use cache
The cache key we restore is keyed on the checked-in go.sum file, however
the cache we save to is keyed on the post-build, modified go.sum. Because
of this, we never actually use the cached modules.

This change causes the cache to be saved under the original go.sum file
rather than the modified one. This still means that some packages will be
downloaded whenever the go.sum is out of date, but that can be resolved by
manually updating go.sum.