.circleci: rework config to improve speed and version coverage
This change reworks the circleci config to accomplish 2 things:
1. dramatically reduce build/test times
2. improve version coverage

For goal 1, this PR has added parallelism and broken up some of the jobs.

For goal 2, We are now testing core with versions 1.13 through 1.17 and
contrib with versions 1.14 through 1.17. New versions can be added simply
by adding them to the list, and old versions can be removed as necessary.