appsec: security events over span tags (#1068)
Implement the new AppSec/APM integration by using the APM service entry spans to report AppSec events.
It results in an important simplification as AppSec no longer needs its own intake client, writer goroutine, event payloads, etc.

The resulting main difference with the previous implementation is that the security event context is now created in the request hot path while it was formerly done in the AppSec writer goroutine. But since this is how the span tag API works right now, I thought it would be better to optimize this in a separate PR when necessary.