profiler: log configuration at profiling start (#1114)
To match what the tracer package currently does, and make debugging a little
easier, log user-supplied configuration when starting a profiler. This is
configurable with WithLogStartup (like tracer) and is on by default.

Tests had to be patched as well since the new, on-by-default logging drowned
out all the test results. Updated tests to discard logs by default for testing
the profiler package.

This also moves osName and osVersion into their own package and exports
them to re-use the code in logging profiler configuration.