tracer.StartSpanFromContext: Fix a race if opts is reused
The opts input slice comes from the caller. If it has extra capacity,
the calls to append will use the slice that is passed in. If this
function is called from multiple goroutines, this causes a race,
where spans may not have the correct parents.

Add a test to demonstrate the error. This test fails with race, and
should also fail without race detection, but that will depend on the
race itself.