internal/appsec/waf: remove appsec rule size limitations (#1189)
The size of the WAF rules were limited like regular WAF values, leading
to truncated WAF rules. This patch changes the limits to the max int
values when encoding the WAF rules. It is considered a simpler change
to achieve compared to the introduction of new "disabled limit" values
that would have introduced more advanced and complex conditions in the
encoder.