internal/traceprof/traceproftest: make sure resource name updates show up in CPU profiles
Updating the resource name for a span after it's created (using SetTag)
should be reflected in the CPU profile.

This test can possibly be flaky due to the nature of sample-based
profiling, so we try to run longer if needed to give a chance for the
desired samples to show up.