profiler: add CPUProfileRate option
If a user wants to use a CPU profiling rate other than the default, they
would need to be able to set the rate before each round of profiling to
avoid having their rate overridden by runtime/pprof.StartCPUProfile
after the first round. This is currently not possible, so add an option
that we can use internally to set the rate each time.

Using this option will cause a spurious warning every time profiling
starts. The warning is annoying, but setting the rate will work as
intended.

Fixes #1242