internal/telemetry: fix flaky client test (#1209)
TestClient was a little over-complicated. It expected the client
messages to come in a specific order. However, it is entirely possible
for the start message, which is sent in a background goroutine, to get
scheduled/show up *after* the heartbeat and stop messages. Make the test
simpler and not scheduler-dependent by just checking that the client
starts up and we get a heartbeat.