internal/log: fix flaky test (#1193)
TestLog/Error/auto waited an arbitrarily short amount of time, assuming
that when the time had elapsed that logs would be flushed. This made the
test flaky on busy CI servers. In reality the test is looking for logs
with the same format in the current logging interval to be suppressed,
so we can just set a very very long interval, do the logs, and call
Flush() to achieve the same result.