internal/appsec/waf: fix the lifetime of waf values (#1245)
We must not free ddwaf_object values passed to the WAF context until
the WAF context gets destroyed at the end of the request.
Lukily, we didn't get any problem because we were calling the context
only once in our HTTP or gRPC integrations.

This patch stops freeing the values and delegates it to the WAF when
it destroys the context. To do so, we need to pass a pointer to the
free function the WAF should use to release the context values to
ddwaf_context_init.

We are still using our own free function here in order to keep making
benefit from the simple allocation monitoring it embeds.