profiler: get the correct profile for goroutine wait profile
PR #842 refactored the profiling code and in the proccess changed the
pprof lookup for the goroutine wait profile from

	pprof.Lookup("goroutine")

to

	pprof.Lookup("goroutinewait")

(by looking up using the profileType.Name field as the key). That
profile type doesn't exist, so the lookup always fails.

The existing TestRunProfile didn't catch this because the test replaces
the real profile lookup with a function that provides a fixed profile.
Add an additional test that runs the real profiles and checks that they
all at least show up in the uploaded batch of profiles.