profiler: pull heap profile C extension into collect function
To maintain the current design where profileType.Collect returns all of
the expected profile data, pull the heap profile special case into the
collectGenericProfile function and modify profileType.Collect to return
the "extra" profile data. That extra parameter is still kind of a
special case since it only exists for the C allocation profile
extension, but at least the special case is in a more logical place.