profiler: pull delta profiling logic into Collect functions
The Collect function will return just the final, serialized profile like
it originally did. This change also simplifies the runProfile function
significantly. In general it's better to have the specific logic for
each profile type more self-contained so it can be changed without
having to change the overall profiler logic.