profiler: add extension hook for C allocation profiling (#1264)
This is a first pass at adding an internal bridge for integrating C
allocation profiling with the regular heap profile. C allocation
profiling will require cgo and possibly other external dependencies and
non-standard linker flags. As such, the main profiler package should not
directly import the C allocation profiler implementation. But there
still needs to be some way to actually _use_ the allocation profiler if
it's available. This commit adds a bridge for doing so, and makes a few
changes to enable actually running the profile and merging it with the
heap profile.