[RCM] Add remote config support to the tracing library
Adds the ability for subsystems running within the tracer to utilize
remote configuration to receive configuration updates within the tracer.

This work primarily relies on the concept of a remote config
`Repository` that was recently introduced into the datadog agent. The
`Repository` is responsible for processing config updates from the core
agent, validating them and keeping an up to date store of the current
configurations of products a client is configured to receive. The code
in the tracer project is mostly a thin networking layer around this
state, as well as code to dispatch config updates to interested
subsystems. (Although at this time with this PR, no subsystems are
actually actively utilizing the client.)

The idea is that when a tracer starts up, if a subsystem that utilizes
remote configuration is enabled, a remote config client will be created
and will start periodically polling the core agent for updates.

This is somewhat of a draft PR at the moment. The tracer was created in
order to help develop system tests that can be used to help validate
tracer clients. Some work is still needed in order to make this
production ready, and some decisions need to still be made about how and
when to configure remote configuration within the tracer.