profiler: add C allocation profiling (behind build tag)
This commit adds C allocation profiling to dd-trace-go behind the
"experimental_cmemprof" build tag.

The profiler/internal/cmemprof implements allocation profiling by
hooking into malloc/calloc/etc. and recording the call stack and
allocation size. Allocations are aggregated and included in the Go
allocation profile that is uploaded to Datadog. The allocation profiler
does not support "live heap" profiling at this time. For full stack
traces, a cgo symbolizer should be used such as
github.com/benesch/cgosymbolizer.

The sampling rate is currently not configurable and is fixed at an
average of one sample per 2MB allocated. This is to avoid committing to
a public API until this has been thoroughly tested on real applications
and is ready to be fully supported.

On Linux, for Go versions <= 1.17, the following environment variable
must be set when building a program with C allocaiton profiling:

	export CGO_LDFLAGS_ALLOW='-Wl,--wrap=.*'

Currently C.malloc is not safe to profile by default. To allow profiling
this function, build with Go >= 1.18 and the following environment
variable:

	export CGO_CFLAGS='-fno-omit-frame-pointer'

This restriction may be lifted in the future.