profiler/internal/cmemprof: do malloc safety check in malloc wrapper
Previously used __builtin_return_address(1) from within
profile_allocation to check whether profiling was called from an unsafe
function. However, passing a non-zero parameter to
__bultin_return_address is not guaranteed to work and would crash the
tests if optimization was enabled (or frame pointers were omitted).

(See https://gcc.gnu.org/onlinedocs/gcc/Return-Address.html)

The only safe way to check the return address is thus to do it from
within the malloc wrapper. Added a second profile_allocation variant
that accepts the return address so we can compute the address in the
malloc wrapper and actually check it only when profiling is enabled and
the allocation would be sampled. This keeps the overhead of the safety
checks low.